Mini-showcase extensions in front page
It uses card to add an animation/screenshot relevant to the extension itself
trying to highlight what it looks like or what's about.

Using a 300x image sometimes is not enough, so there are some images that are
not super great. We can iterate on them to change these later if required.